# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evidencija Recepata API",
        "description": "REST API za upravljanje receptima, sastojcima i korisnicima.",
        "contact": {
            "email": "admin@evidencijarecepata.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Glavni server"
        }
    ],
    "paths": {
...
        "/api/recepti": {
            "get": {
                "tags": [
                    "Recepti"
                ],
                "summary": "Lista svih recepata",
                "operationId": "0f1c5398524ccaa2e5d0d8d866c6437f",
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recepti"
                ],
                "summary": "Kreiranje novog recepta (amater ili profesionalac)",
                "operationId": "79631c89e287bd2abe180ada6ced9235",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "opis",
                                    "kategorija"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Pasulj"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Domaći pasulj sa dimljenim mesom"
                                    },
                                    "kategorija": {
                                        "type": "string",
                                        "example": "supe"
                                    },
                                    "putanja_slike": {
                                        "type": "string",
                                        "example": "slike/pasulj.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recept kreiran"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recepti/{id}": {
            "get": {
                "tags": [
                    "Recepti"
                ],
                "summary": "Detalji jednog recepta sa sastojcima",
                "operationId": "75e134aebbb0cf1ff17770597acae5f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recept nije pronađen"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recepti"
                ],
                "summary": "Izmena recepta (samo vlasnik)",
                "operationId": "489b86514542754ad104de8ce2ae6c17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Novi naziv"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Novi opis"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recept izmenjen"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Recept nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recepti"
                ],
                "summary": "Brisanje recepta (samo vlasnik)",
                "operationId": "058693126d6d931f01489143a832c538",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recept obrisan"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Recept nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recepti/{id}/potvrda": {
            "post": {
                "tags": [
                    "Recepti"
                ],
                "summary": "Potvrda ili odbijanje recepta (samo profesionalac)",
                "operationId": "a0316e308ceeed9a6fc252699fcf14c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "potvrdjen",
                                            "odbijen"
                                        ],
                                        "example": "potvrdjen"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status recepata izmenjen"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Unesi token dobijen pri registraciji/prijavi",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}